fix: resolve CI build errors and clean up API branch structure
1. Fix stego_cmd.cpp sign comparison warning:
   - Cast filename_len to size_t in comparison with vector.size()
   - Resolves -Werror=sign-compare error on Linux GCC and MinGW

2. Refactor Doxygen workflow for clean API branch:
   - Use orphan branch to create clean slate (no source code)
   - Deploy ONLY doxygen/ folder and README.md to API branch
   - Remove all other files from API branch
   - Add root README.md with quick start guide
   - Add doxygen/README.md with contents description

Result: API branch will contain only documentation, not full repo.

+++ src/cli/commands/stego_cmd.cpp
@@ -184,7 +184,7 @@ int StegoCommand::do_extract() {
             uint16_t filename_len = extracted_data[0] | (extracted_data[1] << 8);
             
             // Check if metadata is present and valid
-            if (filename_len > 0 && filename_len < 256 && extracted_data.size() > 2 + filename_len) {
+            if (filename_len > 0 && filename_len < 256 && extracted_data.size() > 2 + static_cast<size_t>(filename_len)) {
                 // Extract filename
                 original_filename.assign(
                     extracted_data.begin() + 2,
